Cleaned up decoding and switched back to using an enum for attribute types, plus many more cleanups
git-svn-id: http://svn.littleshoot.org/svn/littleshoot/trunk/common/stun/stack@645 8d7aff80-242d-0410-bcee-ed6f15385b3d

+++ src/main/java/org/lastbamboo/common/stun/stack/encoder/StunAttributeEncoder.java
@@ -93,7 +93,7 @@ public class StunAttributeEncoder implements StunAttributeVisitor
 
     private void writeHeader(final StunAttribute sa)
         {
-        MinaUtils.putUnsignedShort(m_buf, sa.getAttributeType());
+        MinaUtils.putUnsignedShort(m_buf, sa.getAttributeType().toInt());
         MinaUtils.putUnsignedShort(m_buf, sa.getBodyLength());
         }
     }
